requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==46.0.3
cssselect2==0.7.0
DateTime==5.5
distro==1.9.0
Django==5.1.4
django-allauth==65.13.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-environ==0.11.2
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
fonttools==4.55.3
freetype-py==2.5.1
gunicorn==25.1.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.13.0
joblib==1.5.3
kombu==5.4.2
lxml==6.0.2
mysqlclient==2.2.6
numpy==2.2.0
openai==2.26.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pika==1.3.2
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycairo==1.29.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.11.0
pyHanko==0.34.0
pyhanko-certvalidator==0.30.0
PyJWT==2.10.1
pypdf==6.7.5
pyphen==0.17.0
python-bidi==0.6.7
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.3
razorpay==2.0.0
reportlab==4.4.10
requests==2.32.3
rlPyCairo==0.4.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
svglib==1.6.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zope.interface==7.2
zopfli==0.2.3.post1
